server/requirements.txt
alembic==1.13.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==2.2.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.3
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.0
greenlet==3.1.0
idna==3.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
PyJWT==2.4.0
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.34
stripe==2.58.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==2.2.0
zipp==3.20.1
